Move structs into appropriate places.
