+++ src/stoke/mod.rs
@@ -1,10 +1,8 @@
-use crate::exec::{Interpreter, InterpreterKind};
-use crate::{exec, parity_wasm_utils, solver};
+use crate::{exec, parity_wasm_utils, solver, Algorithm, SuperoptimizerOpts};
 use clap::arg_enum;
 use parity_wasm::elements::{Instruction, Internal, Module};
 use rand::distributions::{Bernoulli, Distribution};
 use rand::Rng;
-use std::path::PathBuf;
 use std::sync::mpsc::channel;
 use structopt::StructOpt;
 
@@ -38,54 +36,6 @@ pub struct StokeOpts {
 
     #[structopt(short, long, default_value = "0.2")]
     pub beta: f64,
-}
-
-#[derive(Clone, Debug, StructOpt)]
-pub enum Algorithm {
-    Stoke(StokeOpts),
-}
-
-#[derive(Clone, Debug, StructOpt)]
-#[structopt(name = "options", about = "Superoptimizer options.")]
-pub struct SuperoptimizerOpts {
-    #[structopt(name = "FILE", parse(from_os_str))]
-    pub input: PathBuf,
-
-    #[structopt(
-        short,
-        long,
-        help="Which interpreter to use for evaluating test cases.",
-        possible_values=&InterpreterKind::variants(),
-        default_value="Wasmer")]
-    pub interpreter_kind: InterpreterKind,
-
-    #[structopt(
-        short,
-        long = "no-opti",
-        help = "If set, run synthesis step only and skip optimization step, true by default.",
-        parse(from_flag = std::ops::Not::not)
-    )]
-    pub opti: bool,
-
-    #[structopt(
-        short,
-        long,
-        help = "The max runtime of one synthesis or optimization step in minutes.",
-        default_value = "5"
-    )]
-    pub time_budget: i64,
-
-    #[structopt(
-        short,
-        long,
-        help = "A comma separated list of integers for initial set of constants.",
-        default_value = "-2,-1,0,1,2",
-        require_delimiter(true)
-    )]
-    pub constants: Vec<i32>,
-
-    #[structopt(subcommand)]
-    pub algorithm: Algorithm,
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
@@ -159,7 +109,7 @@ impl Superoptimizer {
     fn eval_candidate(
         &self,
         mode: Mode,
-        interpreter: &dyn Interpreter,
+        interpreter: &dyn exec::Interpreter,
         candidate: &mut Candidate,
     ) -> u32 {
         let mut cost = if self.stoke_options.enforce_stack_check {
@@ -213,7 +163,7 @@ impl Superoptimizer {
     fn do_run(
         &self,
         mode: Mode,
-        interpreter: &mut dyn Interpreter,
+        interpreter: &mut dyn exec::Interpreter,
         candidate: &mut Candidate,
     ) -> bool {
         let func_type = candidate.spec_func_type();
